+++ tests/unit/test_e2e.py
@@ -100,7 +100,12 @@ GENERIC_ICAL_RESULT_PATH = Path(dir_path, "data", "ical", "ical1_result.json")
         (
             Equinix,
             [("email", Path(dir_path, "data", "equinix", "equinix1.eml"))],
-            [Path(dir_path, "data", "equinix", "equinix3_result.json")],
+            [Path(dir_path, "data", "equinix", "equinix1_result_combined.json")],
+        ),
+        (
+            Equinix,
+            [("email", Path(dir_path, "data", "equinix", "equinix3.eml"))],
+            [Path(dir_path, "data", "equinix", "equinix3_result_combined.json")],
         ),
         # EUNetworks
         (EUNetworks, [("ical", GENERIC_ICAL_DATA_PATH),], [GENERIC_ICAL_RESULT_PATH,],),